edite

+++ controllers/tours/tour_update.js
@@ -7,38 +7,7 @@ const TourUpdate = async(req,res) => {
     // console.log(id);
     // console.log(req.files);
     const{ name,description,image,price,distance,dayData} = req.body;
-
-    // if(req.files){
-    //     let newfile_name = req.files[0].filename;
-    //     console.log(newfile_name);
-    //     if(image==='null'){
-    //         const query1 = `UPDATE tour SET tour_img='${newfile_name}' WHERE tour_id='${id}'`;
-    //         DB.connection.query(query1,(err,result)=>{
-    //             if(result){
-    //                 console.log(result);
-    //             }else{
-    //                 console.log(err);
-    //             }
-    //     })
-    //     }else if(image!=='null'){
-    //         //delete file from storage
-    //         const filePath  = `./uploads/tour/${image}`;
-    //         try {
-    //             await fs.unlink(filePath);
-    //             console.log('File deleted successfully');
-    //             const query1= `UPDATE tour SET tour_img='${newfile_name}' WHERE tour_id='${id}'`;
-    //             DB.connection.query(query1,(err,result)=>{
-    //                 if(result){
-    //                     console.log(result);
-    //                 }else{
-    //                     console.log(err);
-    //                 }
-    //             })
-    //         } catch (error) {
-    //             console.log(error);
-    //         }
-    //     }
-    // }
+    console.log(req.body);
 
     if(name,description,price,distance){
         const query2= `UPDATE tour SET tour_name='${name}',tour_description='${description}',tour_price=${price},distance='${distance}' WHERE tour_id='${id}'`;
@@ -52,8 +21,6 @@ const TourUpdate = async(req,res) => {
     }
 
     if(dayData){
-        // console.log(dayData[0]);
-        // const Data = dayData[0];
         dayData.forEach(async(day,Index)=>{
             console.log(`Day ${day.day} - Date ID: ${day.dateId}`);
 
@@ -77,9 +44,7 @@ const TourUpdate = async(req,res) => {
             })
 
 
-            if(day.places.length>0){
-                console.log(day.dateId)
-                console.log(day.places);
+            if( day.places && day.places.length>0){
 
                 //delete data
                 const query3 = `DELETE FROM tour_places WHERE tour_date_id='${day.dateId}'`;
